DEV: reorganize cluster files for easier development
